[dataflow] Rename to resolve conflicts on some testing.
