+++ python/ray/worker.py
@@ -21,7 +21,7 @@ import ray.services as services
 from ray.services import logger
 
 import ray.dataflow.distributor as distributor
-import ray.dataflow.logging as worker_logger
+import ray.dataflow.logger as worker_logger
 import ray.dataflow.object_store_client as object_store_client
 import ray.signature
